backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "adithep.b",email = "adithepbaebmueann@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13, <4.0"
dependencies = [
    "fastapi[all] (>=0.116.1,<0.117.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "fastapi-cli (>=0.0.8,<0.0.9)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "firebase-admin (>=7.0.0,<8.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "ultralytics (>=8.3.167,<9.0.0)",
    "torch (==2.7.1)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "pyzbar (>=0.1.9,<0.2.0)",
    "easyocr (>=1.7.2,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "datetime (>=5.5,<6.0)",
    "networkx (>=3.5,<4.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "weasyprint (>=65.1,<66.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
